refactor: List 컴포넌트 관련 하나의 폴더로 묶음
